[Fix] renam choice to init
