+++ autoPyTorch/pipeline/traditional_tabular_classification.py
@@ -9,8 +9,8 @@ from sklearn.base import ClassifierMixin
 
 from autoPyTorch.pipeline.base_pipeline import BasePipeline
 from autoPyTorch.pipeline.components.base_choice import autoPyTorchChoice
-from autoPyTorch.pipeline.components.setup.traditional_ml.base_model_choice import ModelChoice
 from autoPyTorch.utils.hyperparameter_search_space_update import HyperparameterSearchSpaceUpdates
+from autoPyTorch.pipeline.components.setup.traditional_ml import ModelChoice
 
 
 class TraditionalTabularClassificationPipeline(ClassifierMixin, BasePipeline):
